// WebApplicationPlateforme/ClientApp/src/app/shared/Models/Dotations/location.model.ts
export class Location {
 id: number;
 numcontrat: string;
 prixlocationmois :string;
 datedebutcontrat :string;
 delaicontrat :string;
 moisdelocation :string;
 locataireName :string;
 idlocataire: number;
 idunite: number;
 iddotation: number;
 nomunite: string;
 nomdotation: string;
 creatorName :string;
 idUserCreator :string;
 dateenreg :string;
}